requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
schedule>=1.2.0
requests>=2.31.0
